Ambient activity context
The purpose of this change is to scope source authentication state to an
abstract activity.

- Introduced an ambient context class with a single property
  'CorrelationId' identifying current activity
- HttpSource makes use of the CorrelationId by mapping it to a
  AuthenticationState, e.g. retries counter, blocking state.
- Added a call to create new activity in key places such as install
  package, uninstall package, restore missing packages, start search
